+++ src/collections/document/document.ts
@@ -26,4 +26,16 @@ export class DocumentCollection extends Collection<Model> {
       this.map(async (model) => await model.toJson(req)),
     );
   }
+
+  /**
+   * Returns Recyclebin JSON data.
+   * @method toRecyclebinJson
+   * @param {Request} req Request object.
+   * @returns {Promise<Json>} JSON object.
+   */
+  async toRecyclebinJson(req: Request): Promise<Json> {
+    return await Promise.all(
+      this.map(async (model) => await model.toRecyclebinJson(req)),
+    );
+  }
 }